Fix issue were node was added to the end
When a node was found to be different, the previous one was removed and
the new one was added. However, this would cause the node to be added to
the end, which lead to incorrect result when it was iterating over the
list of nodes.

This expects the node to have `replaceNode` that allows it to replace in
place