Make it work properly with Godot
There was an issue where when we used the library instead of the SDK,
you would be able to use the generated nodes and components, but not
the inbuilt renderer.

By setting the SDK properly in the library as well as not generating
assembly info and target framework attributes, it is now possible to
include `ReactiveSharpGodot` into your project and start creating
components!