Rewrite the rendering logic
The previous logic did not create a tree out of the components. This
new implementation does. This significantly simplifies the building and
updating of the different components. In addition, it also fixes an
issue which the previous implementation had related to the resetting of
state.

The previous implementation did not have an easy way of resetting state
for components that were dropped when you dropped a grand-parent of it
from the tree. This uses the `Component.Key` to reset all the unused
components.