Use own counter for nodes
When godot names the nodes, it has a counter to make sure that all the
names in the tree are unique. While this is nice for the engine, it does
make things tricky when you are trying to use the names and the tree for
test output.

Instead of using this, we simply override the names of all the nodes
that are created within the test scene and reset this counter between
each test. This makes the count after the name of the node consistent