Only increase counter on unique additions
All nodes added to the tree is kept in a list of weak references. This
allows us to make sure that we only increment the counter on elements
that are new to the tree. Removing and adding the same node should not
make the node get a new name with a higher number