Insert nodes in their correct position
Prior to this change, doing something simple as re-rendering after a
state change with two elements in the tree would cause them to re-order
as one would be removed and re-added at the end.

The position of rendered elements are now stored so they are always
inserted were they are expected to be. This allows better reuse of
nodes.